Fix: Wrap login page component in Suspense boundary
Refactored the login page to resolve a Next.js build error and a related functional bug.

The `useSearchParams` hook was being used in a page component without a Suspense boundary, causing the build to fail. This also led to a UI glitch where the user was incorrectly redirected to the mobile number entry step after OTP submission.

This was fixed by:
1. Extracting the login form logic into a new client component, `LoginComponent`.
2. Replacing the content of the original `login/page.tsx` with a new server component that renders `LoginComponent` inside a `<React.Suspense>` boundary.
3. Adding a `useEffect` cleanup in `LoginComponent` to dispatch `resetOtpSent`, which prevents stale state from disrupting the login flow.

Note: The application's build is currently failing due to an unrelated issue with image imports. This prevented full end-to-end and frontend verification, but the original build error related to `useSearchParams` has been resolved.
