--- dev_server.log
@@ -1,3 +1,0 @@
-
-> alist-dashboard@0.1.0 dev
-> next dev
