+++ src/app/(auth)/login/LoginComponent.tsx
@@ -1,0 +1,117 @@
+"use client";
+
+import { useState, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { RootState, AppDispatch } from '@/store/store';
+import { sendOtpRequest, loginRequest, resetOtpSent } from '@/store/auth/authSlice';
+import Image from 'next/image';
+import { useRouter, useSearchParams } from 'next/navigation';
+import SlideCaptcha from '@/components/general/SlideCaptcha';
+import CountryCodeDropdown from '@/components/general/CountryCodeDropdown';
+
+export default function LoginComponent() {
+  const [phoneNumber, setPhoneNumber] = useState('');
+  const [countryCode, setCountryCode] = useState('+971');
+  const [otp, setOtp] = useState('');
+  const [isCaptchaVerified, setIsCaptchaVerified] = useState(false);
+  const dispatch = useDispatch<AppDispatch>();
+  const { loading, error, otpSent, isAuthenticated, phoneNumber: storedPhoneNumber, loginInProgress } = useSelector((state: RootState) => state.auth);
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const redirectedFrom = searchParams.get("redirectedFrom");
+  const redirectUrl = searchParams.get("redirect");
+
+  useEffect(() => {
+    if (isAuthenticated && !loginInProgress) {
+      if (redirectedFrom === "admin" && redirectUrl) {
+        router.push(redirectUrl);
+      } else {
+        router.push("/dashboard");
+      }
+    }
+    // Cleanup function to reset the otpSent flag when the component unmounts
+    return () => {
+      dispatch(resetOtpSent());
+    };
+  }, [isAuthenticated, loginInProgress, router, dispatch, redirectedFrom, redirectUrl]);
+
+  const handleSendOtp = (e: React.FormEvent) => {
+    e.preventDefault();
+    dispatch(sendOtpRequest({ phoneNumber, country_code: countryCode }));
+  };
+
+  const handleLogin = (e: React.FormEvent) => {
+    e.preventDefault();
+    dispatch(loginRequest({ phoneNumber, otp, country_code: countryCode }));
+  };
+
+  const handleCaptchaSuccess = () => {
+    setIsCaptchaVerified(true);
+  };
+
+  return (
+    <div className="m-auto px-2">
+      <div className="bg-white rounded-[28px] shadow-[0_0_4px_rgba(0,0,0,0.16)] px-[22px] py-[52px] w-[358px] max-w-full">
+        <div className="mb-[47px]">
+          <div className="flex flex-col items-center">
+            <Image
+              src="/icons/AlistLogo.svg"
+              alt="logo"
+              width={100.09}
+              height={109.42}
+            />
+          </div>
+        </div>
+        {!otpSent ? (
+          <form className="w-full flex flex-col gap-[11px]" onSubmit={handleSendOtp}>
+            <div className="flex gap-2">
+              <CountryCodeDropdown selectedCode={countryCode} onCodeChange={setCountryCode} />
+              <input
+                type="text"
+                placeholder="Phone Number"
+                value={phoneNumber}
+                onChange={(e) => {
+                  const numericValue = e.target.value.replace(/[^0-9]/g, "");
+                  setPhoneNumber(numericValue);
+                }}
+                className="rounded-[11px] bg-gray-100 px-5 py-[11px] text-[15px] outline-none focus:ring-2 focus:ring-[#00A4B6] transition placeholder:text-[#6E6E6E] w-full"
+                autoComplete="tel"
+              />
+            </div>
+            <button
+              type="submit"
+              className="mt-[11px] rounded-[11px] bg-[#00A4B6] text-white font-semibold py-[11px] text-[15px] hover:bg-[#0090a6] transition"
+              disabled={loading}
+            >
+              {loading ? 'Sending OTP...' : 'Send OTP'}
+            </button>
+            {error && <p className="text-red-500 text-sm mt-2">{error}</p>}
+          </form>
+        ) : (
+          <form className="w-full flex flex-col gap-[11px]" onSubmit={handleLogin}>
+            <p className="text-center text-gray-600">Enter the OTP sent to {storedPhoneNumber}</p>
+            <input
+              type="password"
+              placeholder="OTP"
+              value={otp}
+              onChange={(e) => setOtp(e.target.value)}
+              className="rounded-[11px] bg-gray-100 px-5 py-[11px] text-[15px] outline-none focus:ring-2 focus:ring-[#00A4B6] transition placeholder:text-[#6E6E6E]"
+              autoComplete="one-time-code"
+            />
+            <div className="mt-4">
+              <SlideCaptcha onSuccess={handleCaptchaSuccess} />
+            </div>
+            <button
+              type="submit"
+              className="mt-[11px] rounded-[11px] bg-[#00A4B6] text-white font-semibold py-[11px] text-[15px] hover:bg-[#0090a6] transition disabled:bg-gray-400"
+              disabled={loading || !isCaptchaVerified}
+            >
+              {loading ? 'Signing in...' : 'Sign In'}
+            </button>
+            {error && <p className="text-red-500 text-sm mt-2">{error}</p>}
+          </form>
+        )}
+      </div>
+    </div>
+  );
+}
